Made program termination Multi-Thread-Safe (MT-Safe)
When SIGTERM or SIGINT is received, variable exitflag should be set to 1
and as a result all threads will exit their infinite loop. Because exitflag
is shared to all threads, accesses to it should be protected by a mutex.
Nevertheless, functions pthread_mutex_{lock,unlock} are not Async-Signal-Safe
(AS-Safe) and cannot be called in the signal handler.
Workaround: The SIGTERM/SIGINT handler sends a SIGUSR1 signal to the main thread
or calls directly the SIGUSR1 handler in case the main thread handles the initial
SIGTERM/SIGINT signal. Then, the SIGUSR1 handler sets variable termination_requested
to 1 and execution returns from the handler. Note that termination_requested variable,
is only accessed by the main thread and no synchronization is required. Finally,
the main thread updates the shared variable exitflag using a mutex, consequently
signaling all threads to terminate.